126039 	124212 	123972 124219 127136

+++ projects/party-master-library/src/lib/shared/components/generic/generic-table/generic-table.component.ts
@@ -293,28 +293,28 @@ export class GenericTableComponent implements OnInit {
   }
   
   onDeleteClick(event: any){
-      // this.onOtherClick.emit($event);
-      // const input =  event.target as HTMLInputElement;
-      let acid = event;
-      let ptype: 'C' | 'V' | 'A' = 'C';
-      if (this.activeRoute === 'vendor') {
-        ptype = 'V';
-      } else if (this.activeRoute === 'general-ledger') {
-        ptype = 'A';
-      }
-      else if(this.activeRoute === 'customer') {
-        ptype = 'C';
-      }
-      this.partyMasterService.onDeleteItem(acid,ptype).subscribe(
-        (response) => {
-          alert('Item deleted successfully.');
-          this.refreshData(ptype);
-         },
-        (error) => {
-          console.error('Error deleting item:', error);
-          alert('Failed to delete the item. Please try again.');
+      if(confirm("Are you sure you want to delete this item?")) {
+        let acid = event;
+        let ptype: 'C' | 'V' | 'A' = 'C';
+        if (this.activeRoute === 'vendor') {
+          ptype = 'V';
+        } else if (this.activeRoute === 'general-ledger') {
+          ptype = 'A';
         }
-      );
+        else if(this.activeRoute === 'customer') {
+          ptype = 'C';
+        }
+        this.partyMasterService.onDeleteItem(acid,ptype).subscribe(
+          (response) => {
+            alert('Item deleted successfully.');
+            this.refreshData(ptype);
+           },
+          (error) => {
+            console.error('Error deleting item:', error);
+            alert('Failed to delete the item. Please try again.');
+          }
+        );
+      }
       
     }
     refreshData(ptype: 'C' | 'V' | 'A') {
@@ -345,8 +345,13 @@ export class GenericTableComponent implements OnInit {
         ptype = 'C';
       }
       this.partyMasterService.onToggleStatus(acid,ptype).subscribe(
-        (response) => {
-          alert('Item toggled successfully.');
+        (response:any) => {
+          if(response.status === "ok"){
+            alert(response.result);
+          }
+          else{
+            alert('Item is inactive.');
+          }
           this.toggleRefresh(ptype);
          },
         (error) => {
